pyproject.toml
[tool.poetry]
name = "notebook"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
jinja2 = "^3.1.3"
python-multipart = "^0.0.9"
pydantic = {extras = ["email"], version = "^2.6.4"}
pymongo = "^4.6.3"
motor = "^3.4.0"
bcrypt = "^4.1.2"
python-jose = "^3.3.0"
pytz = "^2024.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
